# internal/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package db

import (
	"time"
)

type Article struct {
	ID       int64  `json:"id"`
	AuthorID int64  `json:"author_id"`
	Title    string `json:"title"`
	Content  string `json:"content"`
	// -1 Stands for not-checked yet, 0-9 is actual rating
	Rating    int32     `json:"rating"`
	CreatedAt time.Time `json:"created_at"`
}

type Author struct {
	ID          int64  `json:"id"`
	Name        string `json:"name"`
	Description string `json:"description"`
	// -1 Stands for not-checked yet, 0-3 is actual rating
	Rating    int32     `json:"rating"`
	CreatedAt time.Time `json:"created_at"`
}